+++ R/FGIndex.R
@@ -1,0 +1,10 @@
+#' Fear and Greed Index Values
+#'
+#' A xts object with Fear and Greed Index value. It has 689 rows, and 1 colum. Extracted from 2022-02-07 to 2023-12-27
+#'
+#' \describe{
+#'   \item{FGI}{Daily Fear and Greed Index Value}
+#' }
+#'
+#' @family data
+"FGIndex"